Desgin: 카카오맵 마커 이미지 및 디자인 수정

+++ src/stores/plan-store.js
@@ -2,6 +2,7 @@ import { defineStore } from "pinia";
 import { addDateToPlan, deleteDateToPlan, deleteSpotFromPlan, getPlanList } from "@/api/plan";
 import { OK } from "@/constant/status";
 import { useNotification } from "@kyvg/vue3-notification";
+import { reload } from "@/util/custom-router";
 
 const { notify } = useNotification();
 
@@ -17,21 +18,23 @@ export const usePlanStore = defineStore("plan", {
   actions: {
     async handleClickAddDate(id) {
       const { status } = await addDateToPlan(id);
-      if (status === OK) alert("날짜 추가!");
+      if (status === OK) notify({ type: "success", text: "날짜가 추가 되었습니다!" });
+      reload();
     },
     async handleClickDeleteDate(id) {
-      // const flag = confirm("정말 삭제?");
-      // if (!flag) return;
+      const flag = confirm("정말 삭제?");
+      if (!flag) return;
       const { status } = await deleteDateToPlan(id);
-      if (status === OK) alert("날짜 삭제!");
+      if (status === OK) notify({ type: "success", text: "날짜가 삭제 되었습니다!" });
+      reload();
     },
     async handleClickDeleteSpot(id) {
       const flag = window.confirm("정말 삭제 하시겠습니까?");
       if (!flag) return;
-      console.log(id);
       const { status } = await deleteSpotFromPlan(id);
       if (status === OK) {
         notify({ type: "success", text: "여행지가 삭제 되었습니다!" });
+        reload();
       }
     },
   },
